all right, Softmax tests have pass both TensorFlow & Theano, and float32 & 64

+++ ipwxlearn/glue/tensorflow/layers/softmax.py
@@ -69,7 +69,7 @@ def get_output_with_sparse_softmax_crossentropy(layer, labels, inputs=None, **kw
     # compute loss, and compose the real output by applying softmax to the logits.
     if layer.num_units == 2:
         logits_squeeze = tf.squeeze(logits, [-1])
-        loss = tf.nn.sigmoid_cross_entropy_with_logits(logits_squeeze, tf.cast(labels, tf.float64))
+        loss = tf.nn.sigmoid_cross_entropy_with_logits(logits_squeeze, tf.cast(labels, logits_squeeze.dtype))
     else:
         loss = tf.nn.sparse_softmax_cross_entropy_with_logits(logits, labels)
